Reorg v1--tokens not set correctly
