+++ __init__.py
@@ -1,0 +1,3 @@
+"""
+Initialization for main module.
+"""
