+++ functions/imports.py
@@ -1,0 +1,23 @@
+"""
+Imports and dependencies for functions module.
+"""
+
+# Imports
+
+import sys
+import json
+import requests
+
+import time
+from datetime import datetime
+
+import paho.mqtt.client as mqtt
+
+# Definitions
+
+RPC_REQUEST = {
+    "kind": "rpc_request",
+    "args": {
+        "label": ""
+    }
+}
